moved to seperate vm file + pr
